Allow long long immediates when compiling with c++98
This patch disables the `long-long` warning when compiling with
`std=c++98`. This is needed to build natively for AArch32.

We used to disable the `c++11-long-long` warning especially for
OSX. However, this warning is LLVM specific and also equivalent go
`long-long`. We should use the latter to be compiler agnostic.

Change-Id: I1a1a9da8bc94b66ef1741875bbe9b6e3358a1677