Update AArch32 test-generation configuration to reduce the number of tests.
Some assembler tests were *exhaustively* testing the encoding of instructions.
This lead to a few test files taking more than 10 minutes to compile with GCC.
(Clang takes around 15 seconds.)
We reduce the number of tests to a level that is still acceptable, but that also
leads to an acceptable compilation time with gcc.

Before:

    time CXX=clang++ scons all -j40
		898.17s user 19.20s system 1418% cpu 1:04.68 total

    time CXX=g++ scons all -j40
		5275.81s user 53.89s system 742% cpu 11:57.93 total

After:

    time CXX=clang++ scons all -j40
    622.55s user 14.86s system 1319% cpu 48.309 total

    time CXX=g++ scons all -j40
    1239.82s user 27.78s system 2283% cpu 55.517 total

Change-Id: I92c98776c99a35eba2652d837aa4792c0eda71b6