Rewrite `VIXL_CHECK` in terms of `abort`.
The `VIXL_CHECK` macro is meant to be used for runtime checks. However
VIXL is built, the check should always be done. We used `assert` to
implement `VIXL_CHECK`, but an assertion in C can be disabled with the
`NDEBUG` macro, which is also how Android builds VIXL.

This patch also replaces uses of asserts with `VIXL_CHECK`.

Change-Id: I576c5b4910c7dee66e37c3fd9f73ad4edca3f4b7