Don't assume that pointer offsets fit in 32 bits.
In particular, low-level instruction helpers like adr, b and the like should
accept full 64-bit offsets. In all cases, an int can hold the offset, but
narrowing the offset on input weakens the size assertion.

Also note that this breaks backwards-compatibility, but only for some esoteric
use-cases that access the Assembler directly: 'adr(x0, 0)' will now fail to
compile because the call is ambiguous between the int64_t and Label* variants.

Change-Id: I9821d742ff7b8e497e79fdede8af35a1d6325d23