Support getting the address of a literal with ADR
This patch changes the printf implementation in the MacroAssembler to
use ADR instead of LDR to get the address of the format string. The
format string is placed directly in the code stream.

Additionally, Literal<const char*> was changed to simply
StringLiteral. It may be surprising to have a completely different
behaviour for this case.

Change-Id: I15c7acd457eb8ffed056a60cc438b9d7b54a94e1