Improve compiler version detection
This patch adds a matcher for compiler versions in tool/util.py, to be
used in SConstruct. We now have a new `utils.CompilerInformation` class,
that can be created from the path of the compiler. It can then be
compared with a string describing the requirements:

    compiler = util.CompilerInformat("/path/to/c++")

    # Check the compiler is clang.
    if compiler == 'clang':
      # Check its version.
      if compiler == 'clang-3.8':
        # Valid only using clang version 3.8 exactly
      if compiler >= 'clang-3.6':
        # Add options that weren't supported in clang before 3.6

    # Check the compiler is at least gcc 5:
    if compiler >= 'gcc-5'
      # Add new options
    if compiler == 'gcc-4.8':
      # Workaround gcc 4.8

Finally, this patch uses this to make sure warnings about the `override`
keyword are supported.

Change-Id: I5c779ba88d6b476b12b66a86b966efe280ae2f8d