Isolate each generated tests into anonymous namespaces
Generated tests all have declarations and definitions with the sames
names, and in a C-like fashion we've used the `static` keywords to make
sure things are local to their compilation unit.

However, we cannot make a *type* local to its compilation unit in C, as
it doesn't make sense. But in C++, due to the template instantiation
phase, we need to be able to do that, as the compiler could create a
single template instantiation for a type that is defined differently in
different files. It can lead to very confusing bugs. This what the
anonymous namespaces are for, they make both funcions and type
definitions local to them.

So this patch drops the use of C-like `static` locality in favor of a
C++ anonymous namespace.

Change-Id: Ic1b74b0a589c250d9e7bb77d4c4864dababc4596