AArch32: Replace uses of EnsureEmitFor with scopes.
The intention is to catch cases where we generate more code than we have
reserved space for. It is likely that many of these scopes reserve insufficient
size, but they all pass the tests that exist. We still need to review each one,
write tests, and update the sizes as appropriate.

Note that EmissionCheckScope is the proper scope to use, but it isn't
implemented consistently for AArch32 yet. However, it will be easy to replace
these uses of CodeBufferCheckScope with EmissionCheckScope once it is available.

Change-Id: Ie852c3688b2ae5b1b531d531ad1c2357808fb1eb