Generate simple tests for MacroAssembler functions.
The new MacroAssembler tests work like the generated Assembler and Simulator
tests. The assembled code is not executed (so the tests take no inputs).

The tests are intended to provide coverage of all the valid uses of the
MacroAssembler, to ensure that it does not crash for valid inputs. There is
potential to extend the framework later to check the disassembly against a
reference, but the reference would need to be checked manually in many cases, so
this is not implemented yet.

Most of this patch is infrastructure, but some simple tests are included.

Change-Id: I8fdb89426c0202cd8b724fbbe7b6e1c077e85be1