Remove Addw, Movw, Subw, Tbb and Tbh from the MacroAssembler
This patch removes instructions that should not be exposed to the user,
instead, the MacroAssembler can choose to generate them. For instance:

- Use Add/Mov/Sub instead of Addw/Movw/Subw. The latter instructions
  just target a specific encoding, they are for the assembler only.
- Use Switch/Case/End to generate Tbb and Tbh instruction if they are
  available.

This patch also removes references to these instructions in the
macro-assembler's delegates as well as simulator tests for them. Of
course, they are still covered by assembler tests.

Change-Id: I86d606f433aad8cdccba8dd9681ff6d94a9adac3