Add MacroAssembler tests for Rsc and Orn
This patch adds manual tests for the InstructionCondRROp delegate, and
fixes two cases where CodeBufferCheckScope triggers an assertion because
we have generated more instructions that we thought.

Also, this patch limits this delegate to only Orn, Orns, Rsc and
Rscs. Let's give an error to users if they have mis used the other
instructions.

Change-Id: I36ad39679839b81aa2fe8362064239aea3baa861