[tests] Add support for negative MacroAssembler tests.
This adds the 'negative_testing' scons flag, which defaults to 'off'. When the
flag is 'on', VIXL_ABORT, VIXL_ABORT_WITH_MSG and VIXL_CHECK are modified so
that they throw an exception instead of aborting. Negative tests can then catch
these exceptions.

One new build configuration is added to tools/test.py:
  mode=debug std=c++11 negative_testing=on

Change-Id: I97f529318ba633650954ad32fed8acedae2a6d8b