MacroAssembler tests for the CondSizeRROp delegate
This patch adds tests covering most of the delegates'
paths. Specifically, this delegate takes care of translating Add, Adc,
Sub and Sbc. As well as allowing register shifted register operands in
T32.

On top of adding tests for these instructions, this patch identified a
case where we should reserve space for 3 instruction, in case the
immediate is not encodable. Additionally, the delegate did not handle
the setflags version of those instructions.

Change-Id: Ice8459314ab38004afa61729eba4d18f3f788593