Replace VRegisterList::Include with explicit variants.
Previously, it wasn't clear whether `list.Include(d0)` should have been true
when the list included s0 but not s1. Now, the user must call a specific variant
of Include to make this clear.

This patch also fixes the implementation of UseScratchRegisterScope accordingly,
and updates the test to check this behaviour. Further testing would be
beneficial here, but this patch is blocking other work.

Change-Id: I25499b8d3a292df217e3338e13d1817a7d125b45