Remove delegate support for load/store exclusive and pld
This patch removes delegates that were trying to handle the following
instructions:

  - pld
  - pldw
  - pli

  - lda
  - ldab
  - ldaex
  - ldaexb
  - ldaexh
  - ldah
  - ldrex
  - ldrexb
  - ldrexh
  - stl
  - stlb
  - stlh

  - ldaexd
  - ldrexd
  - stlex
  - stlexb
  - stlexh
  - strex
  - strexb
  - strexh

  - stlexd
  - strexd

Instead, the user will get an error if they are used incorrectly. We no
longer try to make them work with addressing modes that are not already
supported by the instruction itself.

Instead of having the delegates use VIXL_ABORT_WITH_MSG directly, I've
changed the `UnimplementedDelegate` method to print the message
"Ill-formed '{}' instruction.". To do this, we now have a
`ToCString(InstructionType)` function to print instructions, defined in
a new file "aarch32/constants-aarch32.cc".

Change-Id: I4e1f955576b88aae3edaf4dad157f11d4755e6ca