Fix very-long-range literal loads.
The code was basically functional but the CodeBufferCheckScope sizes had to be
increased. The generated code could be improved, but its a corner case anyway so
it's not a priority for now. Here's an example:

    0x004223f2      f2af31ec        adr r1, 0x00422008
    0x004223f6      f44f5c00        mov ip, #8192
    0x004223fa      f2c00c42        movt ip, #66
    0x004223fe      eba1010c        sub r1, ip
    0x00422402      6809            ldr r1, [r1]

This sequence occurs for offset ranges over about 1MB. For ranges below 1MB, the
worst case is still adr+sub+ldr.

Change-Id: I6345127aed12abb59c819f63167170d57160da0b