Make test.py build with different values for target_arch
So that we do not triple the time it takes for running test.py, this
patch adds the ability to set build options as
"test_independently = True". This will prevent the script from including
said build option into the product computing the list of scons
commands. Instead we just add them to the list at the end. We now have
"negative_testing" and "target_arch" as independent build options.

As a result, test.py now runs scons an extra 6 times:

- target_arch=aarch32 + clang++
- target_arch=aarch32 + g++
- target_arch=aarch64 + clang++
- target_arch=aarch64 + g++
- negative_testing=on + clang++
- negative_testing=on + g++

Change-Id: I7d53d6d76971477b2fc7686c409f2620500097ec