[aarch32] Dot not abort when one uses unpredictable or strongly discouraged instructions
Calling VIXL_ABORT when a MacroAssembler object would call
AllowUnpredictable or AllowStronglyDiscouraged would prevent us from
handling those cases in delegates. It should be up to a delegate to
decide whether to give the user an error or generate an alternative code
sequence.

Additionally, the user is able (but that's not recommended of course) to
lift the UNPREDICTABLE or STRONGLY DISCOURAGED limitation using
scopes. Since the AllowUnpredictable() and AllowStronglyDiscouraged()
methods were virtual and overriden by the macro-assembler to
abort. There was no way to write the following code:

```c++
MacroAssembler masm(T32);
{
  ExactAssemblyScope scope(&masm, ...);
  AllowUnpredictableScope allow_unpredictable(&masm);
  // Here be dragons!
}
```

This patch fixes this by turning AllowUnpredictable() and
AllowStronglyDiscouraged() into non-virtual methods.

Finally, the related scopes were inside the `aarch32::Assembler::`
namespace, which is inconsistent with other scopes. Move it to
`aarch32`.

Change-Id: I42e5b772e850a8c23521385cf399aaae77b5eab6