EmissionCheckScope: Fix the operation order when closing
On `Open`, `EmissionCheckScope` will first block the pools and then put
a limit on code generation using `CodeBufferCheckScope`. This means that
on `Close`, we should do the opposite in the *opposite* order, which
wasn't the case.

This fixes a crash where a veneer pool is generated on `Release` of the
veneer pool, going above the limit set by the `CodeBufferCheckScope`.

Change-Id: I0b6ea0c97d9858f5000426f29b10145417505a46