Clean up the Label class.
This adds a separate Label class to be used by the Disassembler, which needs
fewer fields than the Assembler, and removes the pc_offset and minus_zero
fields from the Label class used by the Assembler. If the user wants a minus
zero offset in an instruction that supports it, they can achieve that by using
a MemOperand with PC as the base.

Change-Id: I891b03cef5027acf7d3d5e42e5446663d80a92a1