Add specialised W and X Register classes.
These allow for the register type to be determined at compile time, so that the
it can be used for overload resolution.

Note that this does break the API in the case of the ternary operator. The
following pattern will no longer work:

    Register r0 = (condition) ? x0 : w0;

The workaround is to explicitly cast each register:

    Register r0 = (condition) ? Register(x0) : Register(w0);

Some existing overload situation may also fail with the new scheme; in general,
resolutions can be added by simply providing specific WRegister and XRegister
implementations, which typically can fall through to the existing Register
implementation with an explicit cast.

In the future, we expect an extension which will make WRegister and XRegister
POD types.

Change-Id: I6da8e3f87161f23de7c13e3c5142d4fc40532090