Remove Switch implementation from VIXL
The Switch implementation is not well tested, and uses Adr with a label
which we are going to temporarily disallow until we support patching
multiple instructions. The reason this is tricky is that pools can be
generated in between each macro-instruction and this can break Adr,
which breaks Switch.

Let's add it again later when the feature is ready.

Change-Id: I30d6adf89e4ddcebada6aa70ff7b6a5dc9239ad4