Remove Adr(Register, Label*) support from the MacroAssembler
The Adr macro-instruction taking a label should, in theory, be able to
handle a label bound anywhere in the 32-bit address space. However this
is not implemented therefore we shouldn't have it exposed in the
MacroAssembler as it can be misleading. One would naturally expect it to
generate the appropriate sequence for any label.

Let's enable this again later once it is entirely implemented.

Change-Id: I186b5f0ba1be4febb26ca548df2e1c056863e975