Extract FP utility functions from Simulator
Rips out some more utility-like FP functions from the Simulator,
so that they can be accessed by other VIXL components.
This patch only covers the functions necessary to support the
conversion of half precision numbers in to a decimal format,
but others could follow.

Modifications were also made to how we call FPProcessException()
and ReadDN() from within the these functions.
The latter should now be passed in as value of the new
UseDefaultNaN enum and the former as an *optional* pointer to a
boolean, which will be set to true if an exception is thrown.

Change-Id: I2f624288ee4d201900a19fdc24278760ae027827