Initial Support + FMOV [FP16]
This patch adds the basic groundwork for the v8.2 half precision extension.
This includes an F16 wrapper class, the fmov instruction support, tests and some
related trace/logging support.

The trace additions in the simulator are "disabled" as they currently cause the trace tests to fail.
These need to be re-generated before we can enable their half precision support.
This should be handled under a separate patch.

Change-Id: I431558a21649a612ba31f91cc4ab8eba9b657947