Check CPU features in the Assembler.
Now, when the Assembler is asked to assemble an instruction, it asserts (in
debug mode) that the necessary features are available.

This also updates the tests to use no features by default, so that they must be
explicitly requested in the test itself. This provides some assurance that the
Assembler contains reasonable checks.

This patch uses variadic macros, which are a C99/C++11 feature, and I had to
silence a warning in the build system for C++98 builds. However, its use is
restricted to the tests, and we already rely on some other C99 features.

Change-Id: Ifb9846504f98c6d0912f60ff25f219ce2b780755