Use a more robust method to look for the Git root.
In particular, checking for a directory named '.git' fails when 'git-worktree'
is used, since worktrees get a file named '.git', not a directory.

Conveniently, Git commands in worktrees behave superficially as if the worktree
is itself a normal Git repository, so we can simply ask Git for the top level
directory and compare it with the provided path.

Change-Id: I41ef4667ae7121b304b03f8c39ec4c700d1f1c4e