Enable tests for more CPU features.
These features are already supported by VIXL, and they should have been enabled
in the patch that introduced these tests.

The noise in the diff exists because these tests are generated from a script.
With this version, the generation should be made stable, so that future updates
using the same script do not generate this kind of noise.

Change-Id: Idd9c37e642e0c51fdba07f82a41496c9d6927c1d