[sve] Implement basic register support.
This is a temporary implementation of both ZRegister and PRegister, sufficient
to allow initial development of SVE.

These act as a means to experiment with design ideas. If these are successful,
we will attempt to implement similar mechanisms throughout VIXL's register
classes, as far as API continuity (if not strict compatibility) allows.

The intention is to reduce our reliance on run-time, debug-only checks.

- Qualifications ("/Z", "/M" and lane sizes) are implemented using the type
  system, so that we can check correctness at compile time, and so that we
  minimise the dynamic information that we pass around.

- This design is easy to adapt to make the constructors constexpr (with C++11).
  This should fix the "initialisation order fiasco" problems that we've fallen
  foul of in the past.

  I have not actually done this in this patch because it requires C++11, and I
  want to experiment with the top-level CPURegister first.

For example, we can now specify that a function requires a size argument:

    void mla(ZRegister zda, ...);

    ...

    // Then:
    mla(z0.VnB(), ...);  // Ok: 'z0.VnB()' is known to have a lane size.
    mla(z0, ...);        // Compile-time error: 'z0' has no inherent lane size.

    ZRegister foo = z0;  // Strip the static lane size information.
    mla(foo, ...);       // Ok: A run-time assertion checks for a lane size.

The lane size itself is not represented in the type system, because all the
possible combinations would result in an explosion in the number of types (and
therefore overloads).

I expect to refine the balance between compile-time and run-time checking as we
discover what's useful and worth the effort.

Change-Id: I2466985e1e73db41537b0ef891d8bebd1c4cd8ac