Silence -Wmissing-noreturn.
This warning generates a lot of false positives because we often create stubs
and compile-time configurations with things like VIXL_UNIMPLEMENTED(). Properly
satisfying -Wmissing-noreturn requires that we annotate functions at their
_declarations_, which means we need to add #ifdefs to the header files.

Conditional compilation in headers can result in incorrect behaviour when
libvixl.so is built with different flags than the code using it. Disabling this
warning allows us to simplify the headers, and should make VIXL more robust.

Change-Id: I9bb727be1cf1e2e003491701be25c7b27bd4e2a7