[sve] Make PRegister* types derive from PRegister.
With this patch, PRegister itself can now represent any P register, with or
without a qualifier or lane size specifier. This makes it possible to make
functions that should accept two or more of the sub-classes, and will be
important for instructions which have both zeroing and merging forms.

Before this patch, PRegister represented a base P register (like "p0"). There is
now no equivalent for this type, but we can add one (deriving from PRegister) if
it looks useful.

Change-Id: I5d68bfcce10df88c886c1475f8d1e28d764e3589