[sve] Add support for movprfx.
Movprfx operates as part of a pair, so this includes tests with various other
instructions. Most of these tests are disabled, since the other instructions
aren't completely implemented.

In particular, this implementation assumes that it is possible to unambiguously
classify instructions using `instr->Mask(<Foo>FMask) == <Foo>Fixed`. Whilst this
_should_ be true, it isn't in practice, and we need to refactor the decode trees
accordingly. For now, movprfx works in positive cases with instructions that are
implemented.

Change-Id: Icc92abfde51a79cda7e56896da02cbcf1a5659e8