[sve] Add support for INCP and DECP.
This includes the signed and unsigned saturating forms, for both vectors and
scalars.

The MacroAssembler implementation will emit `movprfx` if a non-destructive
operation is requested. Since this is likely to be a common pattern, this patch
introduces a new MacroAssembler-specific scope utility. The implementation is
incomplete, but sufficient for the instructions implemented here.

Change-Id: Ic27c3e24b4f355727ff1ff0d74f4dc88c20aaa98