[sve] Implement most 'SVEPredicateMisc' instructions.
This omits the 'rdffr' variants because we have not yet implemented FFR.

The encodings for this group are very varied so I have split the group.
We will shortly do this for all such groups.

Some existing tests constructed all-true predicates using a loop and
`Initialise`. This patch replaces instances of that pattern with
`Ptrue`.

Change-Id: Id0fa341dab90ff4880fbe8ada17eccd58f0de338