[sve] Fix handling of overflow in {sq,uq}{add,sub}.
The signed saturation instructions in particular have strange semantics
because the immediate form takes an unsigned immediate, but the vector
form treats the lanes as signed quantities.

This patch removes the ability of the MacroAssembler to synthesise
immediates for affected macros.

This patch also fixes simulation of these instructions, in particular
when detecting saturation.

Change-Id: I81dd1e958ad70dcdba0c5b6155bf1bfa9a2634a4