[sve] Fix the behaviour of SVE_MUL_VL.
The "MUL VL" modifier multiplies by the in-memory size of the vector,
rather than the size of a full Z register. This fix updates the tests,
Simulator and MacroAssembler accordingly.

This patch also moves the public `Adr` macro to an internal (protected)
helper function. The interface required (passing vl_divisor) is not
intuitive outside the implementation of a load or store macro.

The (currently unimplemented) scatter-gather form of `Adr` still exists,
but it will not support SVE_MUL_VL, and so doesn't need the context that
makes the contiguous form's API awkward.

Change-Id: Icc50068cd54ce06353fd02554eda784a9dc2ee23