[sve] Make pre-SVE register tracing consistent.
This introduces consistent rules for register tracing, and introduces
clear rules for displaying and parsing structured accesses. This fixes
several problems in the existing system, and provides the framework for
adapting the SVE tracing to match.

In particular, it is now possible to parse individual lines of trace to
build a model of machine state, without having to examine surrounding
lines.

Documentation is provided for the new format. It is not a formal
specification, but should be sufficient for authors of parsing tools.

The SVE tracing will be updated in a follow-up patch.

Change-Id: Id48dbdb28e9371a7d1359bbe17c8f24e2c5fb442