[sve] Fix premature truncation of shift amount
In SVEBitwiseShiftHelper(), we assign the shift amount to an unsigned, losing
the upper 32-bits of the shift. As this value has to be saturated, rather than
truncated, some 64-bit input values would appear as a small shift inside the
normal range, rather than large shifts yet to be saturated.

Fix that and add some comments and a regression test.

Change-Id: Iadc98452c4d6b4a9588a4a7f14ba8dd8e697a6b3