[simulator] Remove instruction instrumentation support.
This code hasn't been used in a long time so let's remove it. If one needs to
extract information from generated code, it's easy to add throw-away code to do
it in the simulator, so having first-class support for this isn't crucial. Also,
being tied to the decoder made the instrumentation support a little too
limiting.

Change-Id: Icf863a8b0bc19ac20bea3eb39e091c5e632982fc