Document that the DecoderVisitor interface is unstable.
This is primarily a documentation change, claiming the ability for us to add
visitors even on minor version updates. This is necessary for adding support for
most ISA extensions.

To mitigate the inconvenience to use-cases that only examine a subset of
visitors, this patch includes a variant of DecoderVisitor that provides default
(empty) visitor implementations.

Change-Id: I42281aaff95b23759868dee67bc3fc18ca8c726f