Handle native-base C64 instructions correctly.
This updates both assembly and disassembly logic, and adds tests for
pre-Morello loads and stores, which have a capability base register in
C64.

The Assembler changes are simple, mainly replacing `GetBaseRegister()`
with `GetBase()`, which returns a CPURegister.

The Disassembler has some new ISA-sensitive base-register placeholders
for native- and alt-base instructions.

Note that VIXL supports several loads and stores that will not co-exist
with Morello. These have not been updated.

Change-Id: I36a7cb8563a7187b2d1b0f850a302c252f83925b