Fix clang build.
Clang warns (and errors due to -Werror) about these comparisons
losing precision as a result of the constants kXMaxInt and kXMaxUInt
being unrepresentable as a double. Although the warning could be
silenced by explicitly casting the constants to doubles, this relies
on implementation-defined rounding behavior so just write out the
floating point literals explicitly.