Use PROT_MAX for proper W^X page allocation.
This is required in order to run Morello tests and examples on CheriBSD.

PROT_MAX is a FreeBSD permission modifier. To correctly implement W^X
pages, used most of VIXL's tests and examples, we must declare to mmap
that we want the ability to make the page executable some time later.

On platforms that don't support PROT_MAX, this has no functional effect.