logic-aarch64: Fix register clearing bug in extractnarrow
Fixes a case in the simulator where the upper lanes of the register will
be cleared before the narrowing operation occurs.

Normally, clearing before hand works... as long as your source and
destination are always different.

When the source and destination are the same this trounces data that the
narrowing operation needs to use.