Morello: Port libvixl.so to purecap.
With this change, libvixl.so builds for Morello purecap (in addition to
hybrid and conventional AArch64). The simulator is not supported on
purecap hosts.

Notably, this separates host from target pointers, and introduces a
`Pointer` abstraction to avoid many #if-guards.

VIXL could now, in principle, be used to write a prototype purecap JIT.

Whilst VIXL builds, the test infrastructure does not yet support purecap
hosts, and most tests fail. Purecap test coverage will come in
subsequent patches, but for now just a few examples are ported, as
proofs of concept.