Morello: Allow StackPointer() to be a capability.
This is a prerequisite for purecap (AAPCS64-cap), which uses csp as a
stack pointer. Unfortunately, this is a breaking change, since
StackPointer() now needs to return a CPURegister.

The usual stack helpers (including the shadow stack mechanism) work with
a capability stack pointer from C64 code. However, Push() and Pop()
won't currently work if an "alt-base" addressing mode would be required,
because Morello lacks alt-base pre- and post-index forms. Further work
could relax this restriction, if required.