Morello: Adapt existing tests to run under purecap.
Insert some translation code into the standard test prologue and
epilogue, so that existing A64 tests can be called from a purecap
environment.

All of the simulator tests are skipped for purecap hosts, as are a few
assembler tests. These need manual porting to work with host
capabilities, or rely on unimplemented features (such as runtime calls).