Add support for implicit checks (#86)
Some runtimes make use of implicit checks, where a potentially
invalid memory access is performed. If the memory access is invalid
then a signal is raised which is then handled by a custom signal
handler which decides how to handle the signal, often returning
execution to a different function.

This patch enables this use case by performing an optional memory
access before any actual memory access. This ensures the signal is
raised in a fixed location (_vixl_internal_AccessMemory) inside the
simulator which allows for signal handlers to determine:

  1. That the signal came from inside the simulator (using
     IsSimulatedMemoryAccess) and can therefore be handled.
  2. That simulation can be resumed at a determined location
     (using GetSignalReturnAddress) by placing that location in
     the signal handlers ucontext instruction pointer register, thus
     allowing execution to continue at the next physical instruction.

If a signal handler correctly handles a signal raised from inside the
simulator then the simulator can return to the main loop to continue
simulation at the program counter. If the signal handler needs to
return to a different function then it can modify the PC via the
existing WritePc function.