Perform implicit checks on store instructions
It is possible for runtimes to perform implicit checks on store
instructions as well as load instructions, therefore support
performing implicit checks on store instructions as well as loads. Do
this by returning true from memory write operations if they succeeded
and false if they failed but were handled by a signal handler.

Implicit checks on store instructions are simulated using a native
memory load because introducing an additional memory store could
change the observable behaviour of multithreaded runtimes. This means
that it is not currently possible to accurately simulate the
behaviour of implicit checks on regions with different access
permissions for reads/writes. For example: if a page has read but not
write permissions then an implicit check would not fail as the
underlying probing instruction will perform a memory read.