Make MOPS tests tolerant of implementation options.
Most MOPS sequences implement one of options "A" or "B", with slightly
different flags in each case. Some `cpy` sequences also permit either
forward or backward iteration, with different register results.

This relies on a new overloaded form of `Equal64` that checks that a
register matches any one of the listed reference values.

In addition, renumber registers to avoid the use of `x16` in a couple of
tests, because this is a MacroAssembler scratch register and could cause
issues in the future.